+++ daily_log/src/view/total/total.css
@@ -20,7 +20,7 @@ body{
 }
 
 #total-back-card{
-    width: 80%;
+    width: 100%;
     height: 100%;
     margin: auto auto 200px;
 }
